admin/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
